Backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-allauth
dj-rest-auth
djangorestframework-simplejwt
psycopg2-binary
Pillow